Added basic frame lowering
Simple function prologue/epilogue and frame index elimination code
added. This currently uses a frame pointer register (register 6, the old
PC register). Might be able to get rid of this, and free it up as a
general-purpose register.

This can now compile a completely empty function that doesn't return
anything. It's progress, of a sort.